Import Weather Observations API, refactor URLs.
